+++ src/index.ts
@@ -1,20 +1,45 @@
-/* istanbul ignore next  */
-const EPSILON = Number.EPSILON || Math.pow(2, -52);
+export interface RofOptions extends Intl.NumberFormatOptions {
+  integerThreshold?: number;
+}
+
+export interface RofResolvedOptions extends Intl.NumberFormatOptions {
+  integerThreshold: number;
+  minimumSignificantDigits: number;
+  maximumSignificantDigits: number;
+}
 
 export class Rof {
-  static isInteger(x: number | string): boolean {
-    x = +x;
-    if (x === 0) return true;
+  private intNumberFormat: Intl.NumberFormat;
+  private decNumberFormat: Intl.NumberFormat;
+  private options: RofResolvedOptions;
 
-    if (x < 1 && x > -1) return false;
+  constructor(locales?: string | string[], options?: RofOptions) {
+    this.options = {
+      minimumSignificantDigits: 2,
+      integerThreshold: Number.EPSILON,
+      style: 'decimal',
+      ...options
+    } as any;
 
-    return Math.round(x) / x - 1 <= EPSILON;
+    if (typeof this.options.maximumSignificantDigits === 'undefined') {
+      this.options.maximumSignificantDigits = this.options.minimumSignificantDigits + 1;
+    }
+
+    const numberFormatOptions: Intl.NumberFormatOptions = {
+      ...options
+    };
+
+    this.intNumberFormat = new Intl.NumberFormat(locales, numberFormatOptions);
+
+    this.decNumberFormat = new Intl.NumberFormat(locales, {
+      minimumFractionDigits: this.options.minimumSignificantDigits,
+      maximumFractionDigits: this.options.minimumSignificantDigits,
+      ...numberFormatOptions
+    });
   }
 
-  constructor(public basePrecision = 2) {}
-
   format(x: number): string {
-    return Rof.isInteger(x) ? this.formatInteger(x) : this.formatDecimal(x);
+    return this.isInteger(x) ? this.formatInteger(x) : this.formatDecimal(x);
   }
   
   formatInteger(x: number): string {
@@ -24,39 +49,36 @@ export class Rof {
       return '-0';
     }
   
-    const localeString  = x.toLocaleString();
+    const localeString  = this.intNumberFormat.format(x);
   
-    return (localeString.length <= (7 + this.basePrecision)) ?
+    return (localeString.length <= (7 + this.options.minimumSignificantDigits)) ?
       localeString :
       this.formatFloat(x);
   }
 
   formatDecimal(x: number): string {
     if (Object.is(x, -0)) {
-      return `-${x.toPrecision(this.basePrecision + 1)}`;
+      return `-${x.toPrecision(this.options.minimumSignificantDigits + 1)}`;
     }
   
     let localeString;
     let N;
   
     if (Math.abs(x) >= 0.4) {
-      localeString = x.toLocaleString(undefined, {
-        minimumFractionDigits: this.basePrecision,
-        maximumFractionDigits: this.basePrecision
-      });
+      localeString = this.decNumberFormat.format(x);
     } else {
       N = this.getRofPrecision(x);
       localeString = x.toPrecision(N);
     }
   
-    return (localeString.length <= (7 + this.basePrecision)) ?
+    return (localeString.length <= (7 + this.options.minimumSignificantDigits)) ?
       localeString :
       this.formatFloat(x, N);
   }
   
   formatFloat(x: number, N?: number): string {
     if (Object.is(x, -0)) {
-      return `-${x.toExponential(this.basePrecision)}`;
+      return `-${x.toExponential(this.options.minimumSignificantDigits)}`;
     }
     N = N || this.getRofPrecision(x);
     return x.toExponential(N - 1);
@@ -73,7 +95,13 @@ export class Rof {
   private getRofPrecision(x: number): number {
     const match = ('' + x).match(/[1-9]/);
     const char = match ? +match[0] : 0;
-    return this.basePrecision + (char < 4 ? 1 : 0);
+    return char < 4 ? this.options.maximumSignificantDigits : this.options.minimumSignificantDigits;
+  }
+
+  private isInteger(x: number): boolean {
+    if (x === 0) return true;
+    if (x < 1 && x > -1) return false;
+    return Number.isInteger(x) || Math.abs(Math.round(x) / x - 1) <= this.options.integerThreshold;
   }
 }